+++ readline/config.log
@@ -10,7 +10,7 @@ generated by GNU Autoconf 2.71.  Invocation command line was
 ## Platform. ##
 ## --------- ##
 
-hostname = car11s4.42barcelona.com
+hostname = car12s5.42barcelona.com
 uname -m = x86_64
 uname -r = 19.6.0
 uname -s = Darwin
@@ -30,8 +30,8 @@ Kernel configured for up to 6 processors.
 Processor type: x86_64h (Intel x86-64h Haswell)
 Processors active: 0 1 2 3 4 5
 Primary memory available: 8.00 gigabytes
-Default processor set: 343 tasks, 1567 threads, 6 processors
-Load average: 2.60, Mach factor: 3.39
+Default processor set: 290 tasks, 1221 threads, 6 processors
+Load average: 1.23, Mach factor: 4.76
 /bin/machine           = unknown
 /usr/bin/oslevel       = unknown
 /bin/universe          = unknown
@@ -1935,7 +1935,7 @@ configure:7074: checking for tgetent
 configure:7074: gcc -o conftest -g -O2   conftest.c  >&5
 Undefined symbols for architecture x86_64:
   "_tgetent", referenced from:
-      _main in conftest-e9e9cc.o
+      _main in conftest-ffd759.o
 ld: symbol(s) not found for architecture x86_64
 clang: error: linker command failed with exit code 1 (use -v to see invocation)
 configure:7074: $? = 1
@@ -2230,7 +2230,7 @@ configure:7421: checking for mbscasecmp
 configure:7421: gcc -o conftest -g -O2   conftest.c  >&5
 Undefined symbols for architecture x86_64:
   "_mbscasecmp", referenced from:
-      _main in conftest-3816a5.o
+      _main in conftest-916b51.o
 ld: symbol(s) not found for architecture x86_64
 clang: error: linker command failed with exit code 1 (use -v to see invocation)
 configure:7421: $? = 1
@@ -2370,7 +2370,7 @@ configure:7428: checking for mbscmp
 configure:7428: gcc -o conftest -g -O2   conftest.c  >&5
 Undefined symbols for architecture x86_64:
   "_mbscmp", referenced from:
-      _main in conftest-6ceca7.o
+      _main in conftest-0c8730.o
 ld: symbol(s) not found for architecture x86_64
 clang: error: linker command failed with exit code 1 (use -v to see invocation)
 configure:7428: $? = 1
@@ -2518,7 +2518,7 @@ configure:7450: checking for mbschr
 configure:7450: gcc -o conftest -g -O2   conftest.c  >&5
 Undefined symbols for architecture x86_64:
   "_mbschr", referenced from:
-      _main in conftest-1e5533.o
+      _main in conftest-8e71e6.o
 ld: symbol(s) not found for architecture x86_64
 clang: error: linker command failed with exit code 1 (use -v to see invocation)
 configure:7450: $? = 1
@@ -2884,7 +2884,7 @@ generated by GNU Autoconf 2.71.  Invocation command line was
   CONFIG_COMMANDS = 
   $ ./config.status 
 
-on car11s4.42barcelona.com
+on car12s5.42barcelona.com
 
 config.status:951: creating Makefile
 config.status:951: creating doc/Makefile
