+++ src/entry/signals.c
@@ -6,7 +6,7 @@
 /*   By: apriego- <apriego-@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 10:50:46 by apriego-          #+#    #+#             */
-/*   Updated: 2023/08/30 12:27:49 by apriego-         ###   ########.fr       */
+/*   Updated: 2023/08/30 19:03:44 by apriego-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
