+++ src/expansor/expansor.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expansor.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: apriego- <apriego-@student.42barcel>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/08/31 13:10:15 by apriego-          #+#    #+#             */
+/*   Updated: 2023/08/31 13:11:33 by apriego-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+// MIRAR QUE SE LE PASSA CON SEGURIDAD
+
+void	expansor(char	*str)
+{
+	
+}
